pyproject.toml
[project]
name = "ai-builder"
version = "1.0.0"
description = "Lightweight CLI framework for building composable AI tools, agents, and pipelines"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "AGPL-3.0-or-later"}
keywords = ["ai", "agents", "rag", "pipeline", "llm", "langchain", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "typer>=0.15",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Built-in tools: document processing
docs = [
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "python-pptx>=1.0",
    "beautifulsoup4>=4.12",
    "striprtf>=0.0.26",
    "openpyxl>=3.1",
    "docx2txt>=0.8",
    "xlrd>=2.0",
    "ebooklib>=0.18",
]
# Built-in tools: embeddings + vector store
rag = [
    "sentence-transformers>=3.3",
    "faiss-cpu>=1.9",
    "langchain-text-splitters>=0.3",
]
# LLM providers
llm = ["openai>=1.0", "anthropic>=0.40"]
# Agent frameworks
agents = [
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-community>=0.3",
    "langgraph>=0.2",
]
# Tracing / observability
tracing = ["langfuse>=2.0"]
# Web dashboard
serve = ["fastapi>=0.115", "uvicorn[standard]>=0.34"]
# Vector DB alternatives
chroma = ["chromadb>=0.5"]
qdrant = ["qdrant-client>=1.12"]
# Web search
search = ["tavily-python>=0.5"]
# Everything
all = [
    "ai-builder[docs,rag,llm,agents,tracing,serve,chroma,qdrant,search]",
]
dev = ["pytest>=8.0", "ruff>=0.8"]

[project.urls]
Homepage = "https://github.com/rohaanuv/ai-builder"
Repository = "https://github.com/rohaanuv/ai-builder"
Issues = "https://github.com/rohaanuv/ai-builder/issues"

[project.scripts]
ai-builder = "ai_builder.cli:app"

[tool.setuptools]
# Flat layout: package files live directly under src/; import name remains ai_builder.
package-dir = {"ai_builder" = "src"}
packages = [
    "ai_builder",
    "ai_builder.commands",
    "ai_builder.core",
    "ai_builder.deploy",
    "ai_builder.templates",
    "ai_builder.tools",
    "ai_builder.tools.document_loader",
    "ai_builder.tools.document_loader.common",
    "ai_builder.tools.document_loader.loader_epub",
    "ai_builder.tools.document_loader.loader_html",
    "ai_builder.tools.document_loader.loader_json",
    "ai_builder.tools.document_loader.loader_pdf",
    "ai_builder.tools.document_loader.loader_rtf",
    "ai_builder.tools.document_loader.loader_slides",
    "ai_builder.tools.document_loader.loader_spreadsheet",
    "ai_builder.tools.document_loader.loader_text",
    "ai_builder.tools.document_loader.loader_word",
    "ai_builder.tools.document_loader.loader_xml",
    "ai_builder.tools.embedder",
    "ai_builder.tools.llm",
    "ai_builder.tools.llm.connectors",
    "ai_builder.tools.retriever",
    "ai_builder.tools.splitter",
    "ai_builder.tools.vector_store",
    "ai_builder.tools.web_search",
    "ai_builder.tracing",
    "ai_builder.visualize",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py313"
[tool.ruff.lint]
select = ["E", "F", "W", "I"]
